+++ api/lib/application/sessions/session-controller.js
@@ -66,20 +66,6 @@ const update = async function (request, h, dependencies = { sessionSerializer })
   return dependencies.sessionSerializer.serialize({ session: updatedSession });
 };
 
-const getAttendanceSheet = async function (request, h, dependencies = { tokenService }) {
-  const sessionId = request.params.id;
-  const token = request.query.accessToken;
-  const i18n = request.i18n;
-
-  const userId = dependencies.tokenService.extractUserId(token);
-
-  const { attendanceSheet, fileName } = await usecases.getAttendanceSheet({ sessionId, userId, i18n });
-  return h
-    .response(attendanceSheet)
-    .header('Content-Type', 'application/pdf')
-    .header('Content-Disposition', `attachment; filename=${fileName}`);
-};
-
 const getSupervisorKitPdf = async function (
   request,
   h,
@@ -408,7 +394,6 @@ const sessionController = {
   getJurySession,
   get,
   update,
-  getAttendanceSheet,
   getSupervisorKitPdf,
   getCandidatesImportSheet,
   getCertificationCandidates,
